refactor: introduce useTxHistoryLoader intermediate hook
